+++ ValveKeyValue/ValveKeyValue.Test/Text/SteamDataTestCase.cs
@@ -1,6 +1,4 @@
-﻿using System.Globalization;
-using NUnit.Framework;
-using ValveKeyValue.Test.Helpers;
+﻿using NUnit.Framework;
 
 namespace ValveKeyValue.Test
 {
@@ -26,17 +24,6 @@ namespace ValveKeyValue.Test
             Assert.That(
                 (string)data["config"][@"2\"],
                 Is.EqualTo(@" TF2_bak;del /q /s bin\*;cd tf;del /q /s *.cache media\*.mov;mkdir download;cd download;move ..\maps .\;move ..\materials .\;move ..\models .\;move ..\particles .\;move ..\resource .\;move ..\sound .\;move ..\scripts .\;cd resource;del game.ico tf.ttf tf2.ttf tf2build.ttf tf2professor.ttf tf2secondary.ttf tfd.ttf;cd ..\sound\ui;del gamestartup*.mp3 tv_tune*.mp3 holiday\gamestartup_*.mp3;:EOF"));
-        }
-        
-        [TestCase(1)]
-        [TestCase(4096)]
-        public void CanHandleBlockingStreams(int maxReadAtOnce)
-        {
-            using var blockingStream = new BlockingStream(maxReadAtOnce);
-
-            KVObject data = KVSerializer.Create(KVSerializationFormat.KeyValues1Text).Deserialize(blockingStream);
-            
-            Assert.That(data["test"].ToInt32(CultureInfo.InvariantCulture), Is.EqualTo(1337));
         }
     }
 }
